requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
fastapi-mail==1.4.1
alembic==1.12.1
psycopg2-binary==2.9.9
boto3==1.34.0
uuid==1.30
